moved around a bunch of files; tried to structure indexing scripts in lib
