--- umich_catalog_indexing/bin/commit
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-port=$1
-HERE=`dirname $0`
-SCRIPTDIR=`realpath $HERE`
-source $SCRIPTDIR/utils.sh
-export SOLR_URL=`solr_url $port`
-
-commit
-
