--- umich_catalog_indexing/bin/delete_all
@@ -1,19 +1,0 @@
-
-require 'httpclient'
-require 'zlib'
-
-url = ENV['SOLR_URL'] + '/update'
-
-unless url
-  raise "SOLR_URL environment variable not defined"
-end
-
-client = HTTPClient.new
-
-begin
-  body = "<delete><query>*:*</query></delete>"
-  client.post(url, body, {'Content-Type' => 'text/xml'})  
-rescue Exception => e
-  puts "Problem deleting: #{e}"
-end
-
