--- umich_catalog_indexing/bin/delete_ids
@@ -1,37 +1,0 @@
-require 'httpclient'
-require 'zlib'
-
-filename = ARGV[0]
-base_url = ARGV[1] || ENV['SOLR_URL']
-
-unless base_url
-  raise "No url passed and SOLR_URL environment variable not defined"
-end
-
-url = base_url + '/update'
-client = HTTPClient.new
-
-total = 0
-begin
-  file = File.open(filename)
-  if /\.gz\Z/.match(filename)
-    file = Zlib::GzipReader.new(file)
-  end
-
-  body = "<delete>"
-  file.each do |l|
-    body << "<id>#{l.chomp}</id>"
-    total += 1
-  end
-  body << "</delete>"
-
-  if total > 0
-    client.post(url, body, {'Content-Type' => 'text/xml'})  
-  else
-    puts "File #{filename} is empty"
-  end
-  puts "Deleted #{total} ids from #{url}\n\n"
-rescue Exception => e
-  puts "Problem deleting: #{e}"
-end
-
