--- umich_catalog_indexing/bin/index_file
@@ -1,60 +1,0 @@
-#!/bin/bash
-
-if [ $# -lt 1 -o $# -gt 3 ] || [[ "$1" == "-h" ]]
-then
-  echo
-  echo "index_file: just index a single file"
-  echo
-  echo "USAGE:"
-  echo "    index_file <filename> <optional_port> <optional_logfile>"
-  echo
-  echo "[\$SOLR_URL will override default core and port ] "
-  echo
-  echo "DEFAULT VALUES"
-  echo "   port: 8025"
-  echo "   logfile: stdout"
-  echo
-  exit
-fi
-
-DEFAULT_WRITER="localhost"
-
-HERE=`dirname $0`
-SCRIPTDIR=`realpath $HERE`
-marcfile=`realpath $1`
-
-# Make sure we're using a valid java
-# and get the utility scripts
-#source $SCRIPTDIR/set_java_home.sh
-
-source $SCRIPTDIR/utils.sh
-
-port=$2
-logfile=$3
-[[ ! -z "$logfile" ]] && logfile=`realpath "$3"`
-
-export target_url=`solr_url $port`
-
-log "\nIndex to $target_url" $logfile
-log "Indexing $marcfile" $logfile
-
-# The whole ridiculous `tee` pipeline is in service of trying to output errors to STDOUT as well as the logfile, so
-# they show up in cron jobs. This might not be the place to do it, though...
-
-if [ -f "$marcfile" ] 
-then
-    export SOLR_URL=$target_url
-    if [[ ! -z $logfile ]]; then
-	$SCRIPTDIR/tindex $DEFAULT_WRITER $marcfile | tee -a $logfile
-    else
-	$SCRIPTDIR/tindex $DEFAULT_WRITER $marcfile 
-    fi  
-else
-  log  "No indexing: Could not find marcfile '$marcfile'" $logfile
-fi
-
-traject_exit_value=$?
-
-commit
-
-exit $traject_exit_value
