--- umich_catalog_indexing/bin/m_delete_ids
@@ -1,27 +1,0 @@
-#!/bin/bash
-
-SCRATCHDIR="/app/scratch"
-
-#set up the file
-pkg_file=$1
-
-SOLR_URL=$2
-
-#copy file from alma-integrations to scratch directory
-# '@' removes printing interactive session
-sftp -oStrictHostKeyChecking=no -oIdentityFile=$SSH_KEY_PATH -b - $ALMA_FILES_USER@$ALMA_FILES_HOST << EOT
-@get $pkg_file scratch
-EOT
-
-#gets just the file name without the parent directory info
-file=${pkg_file##*/}
-#filename without the filename extension
-base=${file%%.*.*}
-tar xzvf scratch/$file -C $SCRATCHDIR
-filename_glob=`echo ${SCRATCHDIR}/${base}.xml`
-
-ls scratch
-
-bundle exec ruby bin/m_delete_ids.rb $filename_glob $SOLR_URL
-
-rm $SCRATCHDIR/${base}*
