--- umich_catalog_indexing/bin/m_get_high_level_browse.rb
@@ -1,11 +1,0 @@
-require "jruby"
-require "high_level_browse"
-path = "./lib/translation_maps"
-
-#does file exist and is file older than one day?
-if !File.exists?("#{path}/hlb.json.gz") or File.stat(path).mtime < Time.now - (60*60*24) 
-  HighLevelBrowse.fetch_and_save(dir: path)
-  puts "updated high level browse"
-else
-  puts "#{path}/hlb.json.gz is less than one day old. Did not update"
-end
