+++ umich_catalog_indexing/lib/jobs.rb
@@ -1,21 +1,17 @@
-require 'sidekiq'
-class IndexIt
-  include Sidekiq::Worker
-  def perform(file, solr_url)
-    puts "indexing #{file} into #{solr_url}"
-    system("bin/mindex_xml #{file} #{solr_url}")
-  end
-end
-class DeleteIt
-  include Sidekiq::Worker
-  def perform(file, solr_url)
-    puts "indexing deletes from #{file} into #{solr_url}"
-    system("bin/m_delete_ids #{file} #{solr_url}")
-  end
-end
-class IndexHathi
-  def perform(file, solr_url)
-    puts "indexing zephir #{file} into #{solr_url}"
-    system("bin/mindex_hathi #{file} #{solr_url}")
-  end
+require 'alma_rest_client'
+
+require "/app/lib/jobs/utilities"
+require "/app/lib/jobs/utilities/sftp"
+require "/app/lib/jobs/utilities/alma_file_processor"
+
+require "/app/lib/jobs/lib_loc_info/libraries"
+require "/app/lib/jobs/lib_loc_info/library_location_list"
+
+require "/app/lib/jobs/delete_id_getter"
+require "/app/lib/jobs/delete_alma_ids"
+
+require "/app/lib/jobs/index_alma_xml"
+require "/app/lib/jobs/index_hathi_json"
+
+module Jobs
 end
