+++ umich_catalog_indexing/lib/jobs/lib_loc_info.rb
@@ -1,0 +1,4 @@
+module Jobs
+  module LibLocInfo
+  end
+end
