+++ umich_catalog_indexing/lib/jobs/utilities.rb
@@ -1,0 +1,4 @@
+module Jobs
+  module Utilities
+  end
+end
